extract element to displaying fetched anime data
element to display animes data got extracted to src/components/AnimeList.tsx to make it more reusable
